Refactor inductor perf dashboard into compoments
