+++ torchci/components/benchmark/CommitPanel.tsx
@@ -1,0 +1,36 @@
+import dayjs from "dayjs";
+import { Stack, Typography } from "@mui/material";
+import { SHA_DISPLAY_LENGTH } from "components/benchmark/common";
+import { BranchAndCommit } from "lib/types";
+import { HUD_PREFIX } from "components/benchmark/common";
+import { ReactNode } from "react";
+
+export function CommitPanel({
+  lBranchAndCommit,
+  rBranchAndCommit,
+  workflowName,
+  children,
+}: {
+  lBranchAndCommit: BranchAndCommit;
+  rBranchAndCommit: BranchAndCommit;
+  workflowName: string;
+  children: ReactNode;
+}) {
+  return (
+    <Stack direction="row" spacing={2} sx={{ mb: 2 }}>
+      <Typography fontSize={"1rem"} fontStyle={"italic"}>
+        *This report was generated by CI running on PyTorch{" "}
+        {lBranchAndCommit.branch} branch at commit{" "}
+        <a href={`${HUD_PREFIX}/${lBranchAndCommit.commit}#${workflowName}`}>
+          {lBranchAndCommit.commit.substring(0, SHA_DISPLAY_LENGTH)}
+        </a>{" "}
+        on {dayjs(lBranchAndCommit.date).format("YYYY/MM/DD")} comparing with{" "}
+        {rBranchAndCommit.branch} branch at commit{" "}
+        <a href={`${HUD_PREFIX}/${rBranchAndCommit.commit}#${workflowName}`}>
+          {rBranchAndCommit.commit.substring(0, SHA_DISPLAY_LENGTH)}
+        </a>
+        . {children}
+      </Typography>
+    </Stack>
+  );
+}
