+++ torchci/components/benchmark/SuitePicker.tsx
@@ -1,0 +1,48 @@
+import {
+  Select,
+  MenuItem,
+  FormControl,
+  InputLabel,
+  SelectChangeEvent,
+} from "@mui/material";
+
+export const SUITES: { [k: string]: string } = {
+  torchbench: "Torchbench",
+  huggingface: "Huggingface",
+  timm_models: "TIMM models",
+  dynamic: "[Dynamic]",
+  blueberries: "[Blueberries]",
+};
+
+export function SuitePicker({
+  suite,
+  setSuite,
+}: {
+  suite: string;
+  setSuite: any;
+}) {
+  function handleChange(e: SelectChangeEvent<string>) {
+    setSuite(e.target.value);
+  }
+
+  return (
+    <>
+      <FormControl>
+        <InputLabel id="suite-picker-input-label">Suite</InputLabel>
+        <Select
+          value={suite}
+          label="Suite"
+          labelId="suite-picker-select-label"
+          onChange={handleChange}
+          id="suite-picker-select"
+        >
+          {Object.keys(SUITES).map((suite) => (
+            <MenuItem key={suite} value={suite}>
+              {SUITES[suite]}
+            </MenuItem>
+          ))}
+        </Select>
+      </FormControl>
+    </>
+  );
+}
