+++ torchci/components/benchmark/common.tsx
@@ -1,0 +1,6 @@
+export const SHA_DISPLAY_LENGTH = 10;
+export const MAIN_BRANCH = "main";
+export const LOG_PREFIX = "https://ossci-raw-job-status.s3.amazonaws.com/log";
+export const HUD_PREFIX = "/pytorch/pytorch/commit";
+export const LAST_N_DAYS = 7;
+export const TIME_FIELD_NAME = "granularity_bucket";
