+++ torchci/package.json
@@ -36,6 +36,7 @@
     "dayjs-plugin-utc": "^0.1.2",
     "echarts": "^5.3.2",
     "echarts-for-react": "^3.0.2",
+    "eslint-plugin-unused-imports": "^3.2.0",
     "jaro-winkler-typescript": "^1.0.1",
     "lodash": "^4.17.21",
     "minimatch": "^9.0.3",
